Calculo de importancia de varias palabras

+++ dlc-trabajo/src/main/java/entities/Documento.java
@@ -36,7 +36,7 @@ public class Documento {
         return valor;
     }
 
-    public void setValor(Integer valor) {
+    public void setValor(float valor) {
         this.valor = valor;
     }
 }
